requirements.txt
Django==1.6.8
selenium==
factory_boy==